+++ lib/lightstreamer.rb
@@ -2,18 +2,18 @@ require 'thor'
 require 'typhoeus'
 require 'uri'
 
+require 'lightstreamer/cli/main'
+require 'lightstreamer/cli/commands/stream_command'
 require 'lightstreamer/control_connection'
 require 'lightstreamer/errors'
 require 'lightstreamer/line_buffer'
+require 'lightstreamer/messages/overflow_message'
+require 'lightstreamer/messages/update_message'
 require 'lightstreamer/session'
 require 'lightstreamer/stream_connection'
 require 'lightstreamer/stream_connection_header'
 require 'lightstreamer/subscription'
-require 'lightstreamer/utf16'
 require 'lightstreamer/version'
-
-require 'lightstreamer/cli/main'
-require 'lightstreamer/cli/commands/stream_command'
 
 # This module contains all the code for the Lightstreamer gem. See `README.md` to get started with using this gem.
 module Lightstreamer